api_design/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
aiohttp==3.9.1
pydantic==2.5.0